+++ apps/frontend/app/(website)/campaigns/[campaignId]/hooks/useRunCodemodMutation.ts
@@ -1,32 +1,26 @@
+import type {
+  CodemodRunBody,
+  CodemodRunResponse,
+} from "@codemod-com/utilities";
 import { RUN_CODEMOD as RUN_CODEMOD_URL } from "@mocks/endpoints/gh-run";
 import { useAPI } from "@studio/hooks/useAPI";
 import { useState } from "react";
 import { useMutation } from "react-query";
 
-// @TODO
-type CodemodRunResponse = {
-  codemodRunIds: Array<{ id: string; codemod: string }>;
-};
-
-type CodemodRunRequest = {
-  codemods: string[];
-  repos: string[];
-};
-
 export const useRunCodemodMutation = () => {
-  const [executionIds, setExecutionIds] = useState<
-    CodemodRunResponse["codemodRunIds"]
-  >([]);
+  const [executionIds, setExecutionIds] = useState<CodemodRunResponse["ids"]>(
+    [],
+  );
 
   const { post: runCodemod } = useAPI<CodemodRunResponse>(RUN_CODEMOD_URL);
 
   const runCodemodMutation = useMutation({
     mutationFn: async (
-      request: CodemodRunRequest,
+      request: CodemodRunBody,
     ): Promise<CodemodRunResponse> => {
       const result = await runCodemod(request);
 
-      setExecutionIds(result.data.codemodRunIds);
+      setExecutionIds(result.data.ids);
 
       return result.data;
     },
