+++ packages/api-types/src/index.ts
@@ -3,3 +3,4 @@ export * from "./responses.js";
 export * from "./github.js";
 export * from "./clerk.js";
 export * from "./campaign.js";
+export * from "./codemod-run.js";
